Replace QPG6100 by QPG6105
