+++ examples/lock-app/APPLICATION.md
@@ -1,11 +1,11 @@
-# Matter QPG6100 Lock Example Application
+# Matter QPG6105 Lock Example Application
 
 An example application showing the use of
-[Matter](https://github.com/project-chip/connectedhomeip) on the Qorvo QPG6100.
+[Matter](https://github.com/project-chip/connectedhomeip) on the Qorvo QPG6105.
 
 ---
 
-- [Matter QPG6100 Lock Example Application](#matter-qpg6100-lock-example-application)
+- [Matter QPG6105 Lock Example Application](#matter-qpg6105-lock-example-application)
   - [Lock-app](#lock-app)
     - [Build and program](#build-and-program)
     - [Button control](#button-control)
@@ -17,15 +17,15 @@ An example application showing the use of
 
 ## Lock-app
 
-The QPG6100 Lock-app shows an implementation of a Lock device, based on
-Matter and the Qorvo QPG6100 SDK. Intention of this example is to demonstrate a Matter
+The QPG6105 Lock-app shows an implementation of a Lock device, based on
+Matter and the Qorvo QPG6105 SDK. Intention of this example is to demonstrate a Matter
 device, with Thread connectivity, using Bluetooth Low Energy to perform Matter provisioning.
 
 ### Build and program
 
 Building and programming instruction can be found on the platform specific page for
 
-- [QPG6100](../../qpg6100/doc/README.md)
+- [QPG6105](../../qpg6105/doc/README.md)
 
 ### Button control
 
@@ -56,7 +56,7 @@ The following LEDs are used during the application:
 
 ### Logging Output
 
-See [View Logging Output](../../qpg6100/doc/README.md#view-logging-output)
+See [View Logging Output](../../qpg6105/doc/README.md#view-logging-output)
 
 At startup you will see:
