--- qpg6100/comps/gpFreeRTOS/config/hooks.c
@@ -1,183 +1,0 @@
-/*
- * FreeRTOS Kernel V10.2.1
- * Copyright (C) 2019 Amazon.com, Inc. or its affiliates.  All Rights Reserved.
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy of
- * this software and associated documentation files (the "Software"), to deal in
- * the Software without restriction, including without limitation the rights to
- * use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of
- * the Software, and to permit persons to whom the Software is furnished to do so,
- * subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all
- * copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS
- * FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR
- * COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER
- * IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
- * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
- *
- * http://www.FreeRTOS.org
- * http://aws.amazon.com/freertos
- *
- * 1 tab == 4 spaces!
- */
-
-/******************************************************************************
- * This project provides two demo applications.  A simple blinky style project,
- * and a more comprehensive test and demo application.  The
- * mainCREATE_SIMPLE_BLINKY_DEMO_ONLY setting (defined in this file) is used to
- * select between the two.  The simply blinky demo is implemented and described
- * in main_blinky.c.  The more comprehensive test and demo application is
- * implemented and described in main_full.c.
- *
- * This file implements the code that is not demo specific, including the
- * hardware setup and FreeRTOS hook functions.
- *
- *
- * Additional code:
- *
- * This demo does not contain a non-kernel interrupt service routine that
- * can be used as an example for application writers to use as a reference.
- * Therefore, the framework of a dummy (not installed) handler is provided
- * in this file.  The dummy function is called Dummy_IRQHandler().  Please
- * ensure to read the comments in the function itself, but more importantly,
- * the notes on the function contained on the documentation page for this demo
- * that is found on the FreeRTOS.org web site.
- */
-
-/* Standard includes. */
-#include <stdio.h>
-
-/* Kernel includes. */
-#include "FreeRTOS.h"
-#include "task.h"
-#include "semphr.h"
-
-#if (configUSE_MALLOC_FAILED_HOOK == 1)
-/*-----------------------------------------------------------*/
-
-void vApplicationMallocFailedHook( void )
-{
-    /* vApplicationMallocFailedHook() will only be called if
-    configUSE_MALLOC_FAILED_HOOK is set to 1 in FreeRTOSConfig.h.  It is a hook
-    function that will get called if a call to pvPortMalloc() fails.
-    pvPortMalloc() is called internally by the kernel whenever a task, queue,
-    timer or semaphore is created.  It is also called by various parts of the
-    demo application.  If heap_1.c or heap_2.c are used, then the size of the
-    heap available to pvPortMalloc() is defined by configTOTAL_HEAP_SIZE in
-    FreeRTOSConfig.h, and the xPortGetFreeHeapSize() API function can be used
-    to query the size of free heap space that remains (although it does not
-    provide information on how the remaining heap might be fragmented). */
-    taskDISABLE_INTERRUPTS();
-    for( ;; );
-}
-#endif
-
-/*-----------------------------------------------------------*/
-#if (configSUPPORT_STATIC_ALLOCATION == 1)
-static StaticTask_t xIdleTaskTCB;
-static StackType_t uxIdleTaskStack[256];
-void vApplicationGetIdleTaskMemory(StaticTask_t **ppxIdleTaskTCBBuffer,
-                                   StackType_t **ppxIdleTaskStackBuffer,
-                                   uint32_t *pulIdleTaskStackSize )
-{
-    /* Pass out a pointer to the StaticTask_t structure in which the Idle tasks
-    state will be stored. */
-    *ppxIdleTaskTCBBuffer = &xIdleTaskTCB;
-
-    /* Pass out the array that will be used as the Idle tasks stack. */
-    *ppxIdleTaskStackBuffer = uxIdleTaskStack;
-
-    /* Pass out the size of the array pointed to by *ppxIdleTaskStackBuffer.
-    Note that, as the array is necessarily of type StackType_t,
-    configMINIMAL_STACK_SIZE is specified in words, not bytes. */
-    *pulIdleTaskStackSize = configMINIMAL_STACK_SIZE;
-}
-#endif
-
-/*-----------------------------------------------------------*/
-#if (configUSE_TIMERS == 1)
-static StaticTask_t xTimerTaskTCB;
-static StackType_t uxTimerTaskStack[configTIMER_TASK_STACK_DEPTH];
-
-void vApplicationGetTimerTaskMemory(StaticTask_t **ppxTimerTaskTCBBuffer,
-                                    StackType_t **ppxTimerTaskStackBuffer,
-                                    uint32_t *pxTimerTaskStackSize)
-{
-    *ppxTimerTaskTCBBuffer = &xTimerTaskTCB;
-    *ppxTimerTaskStackBuffer = uxTimerTaskStack;
-    *pxTimerTaskStackSize = configTIMER_TASK_STACK_DEPTH;
-}
-#endif
-
-/*-----------------------------------------------------------*/
-#if (configUSE_IDLE_HOOK == 1)
-#if defined(__GNUC__)
-__attribute__((weak)) void vApplicationIdleHook(void)
-#elif defined(__IAR_SYSTEMS_ICC__)
-__weak void vApplicationIdleHook(void)
-#else
-#error "Unknown compiler"
-#endif
-{
-    /* vApplicationIdleHook() will only be called if configUSE_IDLE_HOOK is set
-    to 1 in FreeRTOSConfig.h.  It will be called on each iteration of the idle
-    task.  It is essential that code added to this hook function never attempts
-    to block in any way (for example, call xQueueReceive() with a block time
-    specified, or call vTaskDelay()).  If the application makes use of the
-    vTaskDelete() API function (as this demo application does) then it is also
-    important that vApplicationIdleHook() is permitted to return to its calling
-    function, because it is the responsibility of the idle task to clean up
-    memory allocated by the kernel to any task that has since been deleted. */
-}
-#endif
-
-/*-----------------------------------------------------------*/
-#if (configCHECK_FOR_STACK_OVERFLOW > 0)
-void vApplicationStackOverflowHook( TaskHandle_t pxTask, char *pcTaskName )
-{
-    ( void ) pcTaskName;
-    ( void ) pxTask;
-
-    /* Run time stack overflow checking is performed if
-    configCHECK_FOR_STACK_OVERFLOW is defined to 1 or 2.  This hook
-    function is called if a stack overflow is detected. */
-    taskDISABLE_INTERRUPTS();
-    for( ;; );
-}
-#endif
-
-/*-----------------------------------------------------------*/
-#if (configUSE_TICK_HOOK == 1)
-#if defined(__GNUC__)
-__attribute__((weak)) void vApplicationTickHook(void)
-#elif defined(__IAR_SYSTEMS_ICC__)
-__weak void vApplicationTickHook(void)
-#else
-#error "Unknown compiler"
-#endif
-{
-    /* This function will be called by each tick interrupt if
-    configUSE_TICK_HOOK is set to 1 in FreeRTOSConfig.h.  User code can be
-    added here, but the tick hook is called from an interrupt context, so
-    code must not attempt to block, and only the interrupt safe FreeRTOS API
-    functions can be used (those that end in FromISR()). */
-    /* Write to a queue that is in use as part of the queue set demo to
-    demonstrate using queue sets from an ISR. */
-}
-#endif
-
-#if (configGENERATE_RUN_TIME_STATS == 1)
-void vConfigureTimerForRunTimeStats( void )
-{
-  
-}
-#error Add implementation for timer
-uint32_t vGetTimerForRunTimeStats( void )
-{
-    return 0;
-}
-#endif /* configGENERATE_RUN_TIME_STATS */
