--- qpg6100/comps/libmbedtls/trng.c
@@ -1,59 +1,0 @@
-/*
- * Copyright (c) 2020, Qorvo Inc
- *
- * This software is owned by Qorvo Inc
- * and protected under applicable copyright laws.
- * It is delivered under the terms of the license
- * and is intended and supplied for use solely and
- * exclusively with products manufactured by
- * Qorvo Inc.
- *
- *
- * THIS SOFTWARE IS PROVIDED IN AN "AS IS"
- * CONDITION. NO WARRANTIES, WHETHER EXPRESS,
- * IMPLIED OR STATUTORY, INCLUDING, BUT NOT
- * LIMITED TO, IMPLIED WARRANTIES OF
- * MERCHANTABILITY AND FITNESS FOR A
- * PARTICULAR PURPOSE APPLY TO THIS SOFTWARE.
- * QORVO INC. SHALL NOT, IN ANY
- * CIRCUMSTANCES, BE LIABLE FOR SPECIAL,
- * INCIDENTAL OR CONSEQUENTIAL DAMAGES,
- * FOR ANY REASON WHATSOEVER.
- *
- * $Change: 183497 $
- * $DateTime: 2021/10/15 15:12:10 $
- */
-
-/** @file "trng.c"
- *
- *  True random number generator for Qorvo chips
-*/
-
-#if !defined(MBEDTLS_CONFIG_FILE)
-#include "mbedtls/config.h"
-#else
-#include MBEDTLS_CONFIG_FILE
-#endif
-
-#include "qvCHIP.h"
-
-#if defined(MBEDTLS_ENTROPY_HARDWARE_ALT) && defined(MBEDTLS_ENTROPY_C)
-
-#include "mbedtls/entropy.h"
-
-int mbedtls_hardware_poll( void *data,
-                           unsigned char *output,
-                           size_t len,
-                           size_t *olen ) {
-    (void) data;
-
-    if(QV_STATUS_NO_ERROR != qvCHIP_RandomGetDRBG(len, output))
-    {
-      return -1;
-    }
-
-    *olen = len;
-    return 0;
-}
-
-#endif /* MBEDTLS_ENTROPY_HARDWARE_ALT && MBEDTLS_ENTROPY_C */
