--- qpg6100/comps/lwip/arch/perf.h
@@ -1,33 +1,0 @@
-/*
- *    Copyright (c) 2020 Project CHIP Authors
- *    Copyright (c) 2018-2019 Google LLC.
- *
- *    Licensed under the Apache License, Version 2.0 (the "License");
- *    you may not use this file except in compliance with the License.
- *    You may obtain a copy of the License at
- *
- *        http://www.apache.org/licenses/LICENSE-2.0
- *
- *    Unless required by applicable law or agreed to in writing, software
- *    distributed under the License is distributed on an "AS IS" BASIS,
- *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *    See the License for the specific language governing permissions and
- *    limitations under the License.
- */
-
-/*
- *
- *    Description:
- *      This file defines processor-architecture-specific constants,
- *      interfaces and types required for LwIP performance
- *      measurement.
- *
- */
-
-#ifndef CHIP_LWIP_FREERTOS_ARCH_PERF_H
-#define CHIP_LWIP_FREERTOS_ARCH_PERF_H
-
-#define PERF_START
-#define PERF_STOP(s)
-
-#endif /* CHIP_LWIP_FREERTOS_ARCH_PERF_H */
